requirements.txt
kopf
kubernetes
PyYAML
